Everything quasi-working, 1d tester updated
The MDB communicates over SPI with the UNO, sets different duty cycles,
calculates speed and direction. AHRS is also talking to the UNO on the
same SPI lines (except slave select), and it works as long as you talk
to the MDB before the AHRS.

Updated the 1-d faulhaber motor test rig ( the one that looks like a
basket) to allow easy removal of the layers. Should be ready for
cutting.